+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -11,4 +11,5 @@ object MetroDirectives : SimpleDirectivesContainer() {
 
   // Dependency directives.
   val WITH_ANVIL by directive("Add Anvil as dependency and configure custom annotations.")
+  val WITH_DAGGER by directive("Add Dagger as dependency and configure custom annotations.")
 }