+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/ProvidesTransformerTest.kt
@@ -330,7 +330,7 @@ class ProvidesTransformerTest : MetroCompilerTest() {
 
     result.assertDiagnostics(
       """
-        e: ExampleGraph.kt:9:22 `@Provides` functions may not be extension functions. Use `@Binds` instead for these.
+        e: ExampleGraph.kt:9:22 `@Provides` functions may not be extension functions. Use `@Binds` instead for these. See https://zacsweers.github.io/metro/bindings/#binds for more information.
       """
         .trimIndent()
     )